Move API files into separate directory.
